phidata-resume-parser/requirements.txt
pdfplumber
python-docx
pandas
beautifulsoup4
python-dotenv
phidata
streamlit
